+++ src/Widgets/MetaTags.php
@@ -112,8 +112,7 @@ class MetaTags extends Widget
     {
         foreach ($this->languages as $language) {
             Yii::$app->getI18n()->callback($language, function () use ($language): void {
-                if ($route = $this->model->getRoute()) {
-                    $url = $this->urlManager->createAbsoluteUrl($route);
+                if ($url = $this->model->getUrl(true)) {
                     $this->view->registerHrefLangLinkTag($language, $url);
                 }
             });
@@ -131,8 +130,8 @@ class MetaTags extends Widget
 
     protected function registerCanonicalUrlTags(): void
     {
-        if ($route = $this->model->getRoute()) {
-            $this->view->registerCanonicalTag($this->urlManager->createAbsoluteUrl($route));
+        if ($url = $this->model->getUrl(true)) {
+            $this->view->registerCanonicalTag($url);
         }
     }